+++ footer.php
@@ -7,7 +7,7 @@
  * @Author: Roni Laukkarinen
  * @Date: 2020-05-11 13:33:49
  * @Last Modified by:   Roni Laukkarinen
- * @Last Modified time: 2022-08-25 14:16:49
+ * @Last Modified time: 2022-09-07 11:57:45
  *
  * @link https://developer.wordpress.org/themes/basics/template-files/#template-partials
  * @package air-light
@@ -19,7 +19,7 @@ namespace Air_Light;
 
 </div><!-- #content -->
 
-<footer id="colophon" class="site-footer has-dark-bg">
+<footer id="colophon" class="site-footer">
 
   <!-- Default footer content that you can safely remove from here... -->
   <div class="site-info">