+++ single.php
@@ -4,7 +4,7 @@
  *
  * @Date:   2019-10-15 12:30:02
  * @Last Modified by:   Roni Laukkarinen
- * @Last Modified time: 2022-02-08 17:03:18
+ * @Last Modified time: 2022-09-07 11:57:39
  *
  * @package air-light
  * @link https://developer.wordpress.org/themes/basics/template-hierarchy/#single-post
@@ -17,7 +17,7 @@ get_header(); ?>
 
 <main class="site-main">
 
-  <section class="block block-single has-light-bg">
+  <section class="block block-single">
     <article class="article-content">
 
       <h1><?php the_title(); ?></h1>